+++ packages/ui/src/helpers/index.ts
@@ -1,2 +1,1 @@
-export * from './auth';
-export * from './utils';
+export * from './authenticator';
